requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.0.0
requests>=2.25.0
psycopg2-binary>=2.9.0
